Minor refactor of coverage report writer.
  * Move out of the main cli package.
  * Extract inner classes.
