+++ runtime/src/main/java/dev/ionfusion/fusion/cli/cover/EstimatedCoverageInfoPair.java
@@ -1,0 +1,18 @@
+// Copyright Ion Fusion contributors. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package dev.ionfusion.fusion.cli.cover;
+
+import dev.ionfusion.fusion._private.HtmlWriter;
+import java.io.IOException;
+
+class EstimatedCoverageInfoPair
+    extends CoverageInfoPair
+{
+    @Override
+    void renderTotal(HtmlWriter html)
+        throws IOException
+    {
+        html.append("???");
+    }
+}
